migrate lint to action
