+++ apps/lockfile-explorer/src/cli/lint/LintCommandLineParser.ts
@@ -4,6 +4,7 @@
 import { ConsoleTerminalProvider, type ITerminal, Terminal } from '@rushstack/terminal';
 import { CommandLineParser } from '@rushstack/ts-command-line';
 import { InitAction } from './actions/InitAction';
+import { LintAction } from './actions/LintAction';
 
 const LINT_TOOL_FILENAME: 'lockfile-lint' = 'lockfile-lint';
 
@@ -25,5 +26,6 @@ export class LintCommandLineParser extends CommandLineParser {
 
   private _populateActions(): void {
     this.addAction(new InitAction(this));
+    this.addAction(new LintAction(this));
   }
 }
